+++ packages/@storefront/flux-capacitor/src/core/reducers/is-fetching.ts
@@ -2,7 +2,9 @@ import Actions from '../actions';
 import Store from '../store';
 
 export type Action = Actions.FetchMoreRefinements
+  | Actions.FetchMorePastPurchaseRefinements
   | Actions.ReceiveMoreRefinements
+  | Actions.ReceiveMorePastPurchaseRefinements
   | Actions.FetchMoreProducts
   | Actions.ReceiveMoreProducts
   | Actions.FetchProductsWithoutHistory
@@ -29,7 +31,9 @@ export const DEFAULT_FETCHING = {
 export default function updateIsFetching(state: State = DEFAULT_FETCHING, action: Action): State {
   switch (action.type) {
     case Actions.FETCH_MORE_REFINEMENTS: return startFetching(state, 'moreRefinements');
+    case Actions.FETCH_MORE_PAST_PURCHASE_REFINEMENTS: return startFetching(state, 'moreRefinements');
     case Actions.RECEIVE_MORE_REFINEMENTS: return doneFetching(state, 'moreRefinements');
+    case Actions.RECEIVE_MORE_PAST_PURCHASE_REFINEMENTS: return doneFetching(state, 'moreRefinements');
     case Actions.FETCH_MORE_PRODUCTS: return startFetching(state, 'moreProducts');
     case Actions.RECEIVE_MORE_PRODUCTS: return doneFetching(state, 'moreProducts');
     case Actions.FETCH_PRODUCTS_WITHOUT_HISTORY: return startFetching(state, 'search');