+++ packages/@storefront/flux-capacitor/src/core/requests/utils.ts
@@ -54,12 +54,12 @@ namespace RequestHelpers {
   export const pastPurchaseProducts: BuildFunction<Partial<Request>, Request> = (state, overrideRequest = {}) => {
     const request: Partial<Request> = {
       ...RequestHelpers.search(state),
+      ...PastPurchaseAdapter.biasSkus(state),
       pageSize: Selectors.pastPurchasePageSize(state),
       query: Selectors.pastPurchaseQuery(state),
       refinements: Selectors.pastPurchaseSelectedRefinements(state),
       skip: Selectors.pastPurchasePageSize(state) * (Selectors.pastPurchasePage(state) - 1),
-      // no sort needed, saves backend from processing this
-      sort: undefined,
+      sort: RequestAdapter.extractSort(Selectors.pastPurchaseSortSelected(state)),
     };
 
     return <Request>{ ...request, ...overrideRequest };